Update day2 solution to verify solutions from shortlist.
The first solution didn't check that the differing characters were in
the same position, so this adds that check. First we check for all pairs
of words that have differing letter counts of 1, narrowing the search
space to only 2 possible options in my case. Then we check
character-by-character to ensure there is only 1 character different,
and if so, that is the final answer.